Revert "Merge remote-tracking branch 'upstream/develop' into develop"
This reverts commit ce5cc0cdb5c93ce09cfb6d81e02ece42a3e0b946, reversing
changes made to 3148ca87ee1d6b2b14eb9e2c3f4949e4be6fd59f.

+++ hi_scripting/scripting/api/FixLayoutObjects.cpp
@@ -737,7 +737,7 @@ bool Array::copy(String propertyName, var target)
 
 int Array::size() const
 {
-	return (int)numElements;
+	return numElements;
 }
 
 bool Array::contains(var obj) const
